Refactor: Extract methods, doc
